@@ -1,18 +1,18 @@
 name: live4tsdk
-description: A starting point for Dart libraries or applications.
+description: A live4t SDK
 # version: 1.0.0
 # homepage: https://www.example.com
 environment:
-  sdk: ">=2.10.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   dartz: ^0.10.0-nullsafety.1
   dio: ^4.0.0-prev1
   freezed_annotation: ^0.14.1
-  json_serializable: ^3.5.1
+  json_serializable: ^4.0.1
 
 dev_dependencies:
-  build_runner: ^1.12.1
+  build_runner: null
   freezed: ^0.14.1
   json_serializable_generator: ^1.0.1
   pedantic: ^1.9.0
