@@ -6,18 +6,18 @@ import 'package:live4tsdk/src/domain/auth/value_objects.dart';
 
 abstract class IAccountRepository {
   Future<Either<AccountFailure, Unit>> createAccount({
-    @required Account account,
+    required Account account,
   });
 
   Future<Either<AccountFailure, Unit>> updateAccount({
-    @required Account account,
+    required Account account,
   });
 
   Future<Either<AccountFailure, Account>> getAccount();
 
   Future<Either<AccountFailure, Unit>> changePassword({
-    @required Account account,
-    @required Password oldPassword,
-    @required Password newPassword,
+    required Account account,
+    required Password oldPassword,
+    required Password newPassword,
   });
 }
