@@ -5,26 +5,26 @@ part 'failures.freezed.dart';
 @freezed
 abstract class ValueFailure<T> with _$ValueFailure<T> {
   const factory ValueFailure.exceedingLength({
-    @required T failedValue,
-    @required int max,
+    required T failedValue,
+    required int max,
   }) = ExceedingLength<T>;
   const factory ValueFailure.empty({
-    @required T failedValue,
+    required T failedValue,
   }) = Empty<T>;
   const factory ValueFailure.multiline({
-    @required T failedValue,
+    required T failedValue,
   }) = Multiline<T>;
   const factory ValueFailure.listTooLong({
-    @required T failedValue,
-    @required int max,
+    required T failedValue,
+    required int max,
   }) = ListTooLong<T>;
   const factory ValueFailure.invalidEmail({
-    @required T failedValue,
+    required T failedValue,
   }) = InvalidEmail<T>;
   const factory ValueFailure.shortPassword({
-    @required T failedValue,
+    required T failedValue,
   }) = ShortPassword<T>;
   const factory ValueFailure.invalidValue({
-    @required T failedValue,
+    required T failedValue,
   }) = InvalidValue<T>;
 }
