@@ -1,5 +1,5 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
-// ignore_for_file: deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies
+// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides
 
 part of 'failures.dart';
 
@@ -9,57 +9,53 @@ part of 'failures.dart';
 
 T _$identity<T>(T value) => value;
 
+final _privateConstructorUsedError = UnsupportedError(
+    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more informations: https://github.com/rrousselGit/freezed#custom-getters-and-methods');
+
 /// @nodoc
 class _$ValueFailureTearOff {
   const _$ValueFailureTearOff();
 
-// ignore: unused_element
   ExceedingLength<T> exceedingLength<T>(
-      {@required T failedValue, @required int max}) {
+      {required T failedValue, required int max}) {
     return ExceedingLength<T>(
       failedValue: failedValue,
       max: max,
     );
   }
 
-// ignore: unused_element
-  Empty<T> empty<T>({@required T failedValue}) {
+  Empty<T> empty<T>({required T failedValue}) {
     return Empty<T>(
       failedValue: failedValue,
     );
   }
 
-// ignore: unused_element
-  Multiline<T> multiline<T>({@required T failedValue}) {
+  Multiline<T> multiline<T>({required T failedValue}) {
     return Multiline<T>(
       failedValue: failedValue,
     );
   }
 
-// ignore: unused_element
-  ListTooLong<T> listTooLong<T>({@required T failedValue, @required int max}) {
+  ListTooLong<T> listTooLong<T>({required T failedValue, required int max}) {
     return ListTooLong<T>(
       failedValue: failedValue,
       max: max,
     );
   }
 
-// ignore: unused_element
-  InvalidEmail<T> invalidEmail<T>({@required T failedValue}) {
+  InvalidEmail<T> invalidEmail<T>({required T failedValue}) {
     return InvalidEmail<T>(
       failedValue: failedValue,
     );
   }
 
-// ignore: unused_element
-  ShortPassword<T> shortPassword<T>({@required T failedValue}) {
+  ShortPassword<T> shortPassword<T>({required T failedValue}) {
     return ShortPassword<T>(
       failedValue: failedValue,
     );
   }
 
-// ignore: unused_element
-  InvalidValue<T> invalidValue<T>({@required T failedValue}) {
+  InvalidValue<T> invalidValue<T>({required T failedValue}) {
     return InvalidValue<T>(
       failedValue: failedValue,
     );
@@ -67,58 +63,62 @@ class _$ValueFailureTearOff {
 }
 
 /// @nodoc
-// ignore: unused_element
 const $ValueFailure = _$ValueFailureTearOff();
 
 /// @nodoc
 mixin _$ValueFailure<T> {
-  T get failedValue;
+  T get failedValue => throw _privateConstructorUsedError;
 
   @optionalTypeArgs
-  TResult when<TResult extends Object>({
-    @required TResult exceedingLength(T failedValue, int max),
-    @required TResult empty(T failedValue),
-    @required TResult multiline(T failedValue),
-    @required TResult listTooLong(T failedValue, int max),
-    @required TResult invalidEmail(T failedValue),
-    @required TResult shortPassword(T failedValue),
-    @required TResult invalidValue(T failedValue),
-  });
+  TResult when<TResult extends Object?>({
+    required TResult Function(T failedValue, int max) exceedingLength,
+    required TResult Function(T failedValue) empty,
+    required TResult Function(T failedValue) multiline,
+    required TResult Function(T failedValue, int max) listTooLong,
+    required TResult Function(T failedValue) invalidEmail,
+    required TResult Function(T failedValue) shortPassword,
+    required TResult Function(T failedValue) invalidValue,
+  }) =>
+      throw _privateConstructorUsedError;
   @optionalTypeArgs
-  TResult maybeWhen<TResult extends Object>({
-    TResult exceedingLength(T failedValue, int max),
-    TResult empty(T failedValue),
-    TResult multiline(T failedValue),
-    TResult listTooLong(T failedValue, int max),
-    TResult invalidEmail(T failedValue),
-    TResult shortPassword(T failedValue),
-    TResult invalidValue(T failedValue),
-    @required TResult orElse(),
-  });
+  TResult maybeWhen<TResult extends Object?>({
+    TResult Function(T failedValue, int max)? exceedingLength,
+    TResult Function(T failedValue)? empty,
+    TResult Function(T failedValue)? multiline,
+    TResult Function(T failedValue, int max)? listTooLong,
+    TResult Function(T failedValue)? invalidEmail,
+    TResult Function(T failedValue)? shortPassword,
+    TResult Function(T failedValue)? invalidValue,
+    required TResult orElse(),
+  }) =>
+      throw _privateConstructorUsedError;
   @optionalTypeArgs
-  TResult map<TResult extends Object>({
-    @required TResult exceedingLength(ExceedingLength<T> value),
-    @required TResult empty(Empty<T> value),
-    @required TResult multiline(Multiline<T> value),
-    @required TResult listTooLong(ListTooLong<T> value),
-    @required TResult invalidEmail(InvalidEmail<T> value),
-    @required TResult shortPassword(ShortPassword<T> value),
-    @required TResult invalidValue(InvalidValue<T> value),
-  });
+  TResult map<TResult extends Object?>({
+    required TResult Function(ExceedingLength<T> value) exceedingLength,
+    required TResult Function(Empty<T> value) empty,
+    required TResult Function(Multiline<T> value) multiline,
+    required TResult Function(ListTooLong<T> value) listTooLong,
+    required TResult Function(InvalidEmail<T> value) invalidEmail,
+    required TResult Function(ShortPassword<T> value) shortPassword,
+    required TResult Function(InvalidValue<T> value) invalidValue,
+  }) =>
+      throw _privateConstructorUsedError;
   @optionalTypeArgs
-  TResult maybeMap<TResult extends Object>({
-    TResult exceedingLength(ExceedingLength<T> value),
-    TResult empty(Empty<T> value),
-    TResult multiline(Multiline<T> value),
-    TResult listTooLong(ListTooLong<T> value),
-    TResult invalidEmail(InvalidEmail<T> value),
-    TResult shortPassword(ShortPassword<T> value),
-    TResult invalidValue(InvalidValue<T> value),
-    @required TResult orElse(),
-  });
+  TResult maybeMap<TResult extends Object?>({
+    TResult Function(ExceedingLength<T> value)? exceedingLength,
+    TResult Function(Empty<T> value)? empty,
+    TResult Function(Multiline<T> value)? multiline,
+    TResult Function(ListTooLong<T> value)? listTooLong,
+    TResult Function(InvalidEmail<T> value)? invalidEmail,
+    TResult Function(ShortPassword<T> value)? shortPassword,
+    TResult Function(InvalidValue<T> value)? invalidValue,
+    required TResult orElse(),
+  }) =>
+      throw _privateConstructorUsedError;
 
   @JsonKey(ignore: true)
-  $ValueFailureCopyWith<T, ValueFailure<T>> get copyWith;
+  $ValueFailureCopyWith<T, ValueFailure<T>> get copyWith =>
+      throw _privateConstructorUsedError;
 }
 
 /// @nodoc
@@ -140,11 +140,13 @@ class _$ValueFailureCopyWithImpl<T, $Res>
 
   @override
   $Res call({
-    Object failedValue = freezed,
+    Object? failedValue = freezed,
   }) {
     return _then(_value.copyWith(
-      failedValue:
-          failedValue == freezed ? _value.failedValue : failedValue as T,
+      failedValue: failedValue == freezed
+          ? _value.failedValue
+          : failedValue // ignore: cast_nullable_to_non_nullable
+              as T,
     ));
   }
 }
@@ -172,22 +174,25 @@ class _$ExceedingLengthCopyWithImpl<T, $Res>
 
   @override
   $Res call({
-    Object failedValue = freezed,
-    Object max = freezed,
+    Object? failedValue = freezed,
+    Object? max = freezed,
   }) {
     return _then(ExceedingLength<T>(
-      failedValue:
-          failedValue == freezed ? _value.failedValue : failedValue as T,
-      max: max == freezed ? _value.max : max as int,
+      failedValue: failedValue == freezed
+          ? _value.failedValue
+          : failedValue // ignore: cast_nullable_to_non_nullable
+              as T,
+      max: max == freezed
+          ? _value.max
+          : max // ignore: cast_nullable_to_non_nullable
+              as int,
     ));
   }
 }
 
 /// @nodoc
 class _$ExceedingLength<T> implements ExceedingLength<T> {
-  const _$ExceedingLength({@required this.failedValue, @required this.max})
-      : assert(failedValue != null),
-        assert(max != null);
+  const _$ExceedingLength({required this.failedValue, required this.max});
 
   @override
   final T failedValue;
@@ -223,38 +228,30 @@ class _$ExceedingLength<T> implements ExceedingLength<T> {
 
   @override
   @optionalTypeArgs
-  TResult when<TResult extends Object>({
-    @required TResult exceedingLength(T failedValue, int max),
-    @required TResult empty(T failedValue),
-    @required TResult multiline(T failedValue),
-    @required TResult listTooLong(T failedValue, int max),
-    @required TResult invalidEmail(T failedValue),
-    @required TResult shortPassword(T failedValue),
-    @required TResult invalidValue(T failedValue),
+  TResult when<TResult extends Object?>({
+    required TResult Function(T failedValue, int max) exceedingLength,
+    required TResult Function(T failedValue) empty,
+    required TResult Function(T failedValue) multiline,
+    required TResult Function(T failedValue, int max) listTooLong,
+    required TResult Function(T failedValue) invalidEmail,
+    required TResult Function(T failedValue) shortPassword,
+    required TResult Function(T failedValue) invalidValue,
   }) {
-    assert(exceedingLength != null);
-    assert(empty != null);
-    assert(multiline != null);
-    assert(listTooLong != null);
-    assert(invalidEmail != null);
-    assert(shortPassword != null);
-    assert(invalidValue != null);
     return exceedingLength(failedValue, max);
   }
 
   @override
   @optionalTypeArgs
-  TResult maybeWhen<TResult extends Object>({
-    TResult exceedingLength(T failedValue, int max),
-    TResult empty(T failedValue),
-    TResult multiline(T failedValue),
-    TResult listTooLong(T failedValue, int max),
-    TResult invalidEmail(T failedValue),
-    TResult shortPassword(T failedValue),
-    TResult invalidValue(T failedValue),
-    @required TResult orElse(),
+  TResult maybeWhen<TResult extends Object?>({
+    TResult Function(T failedValue, int max)? exceedingLength,
+    TResult Function(T failedValue)? empty,
+    TResult Function(T failedValue)? multiline,
+    TResult Function(T failedValue, int max)? listTooLong,
+    TResult Function(T failedValue)? invalidEmail,
+    TResult Function(T failedValue)? shortPassword,
+    TResult Function(T failedValue)? invalidValue,
+    required TResult orElse(),
   }) {
-    assert(orElse != null);
     if (exceedingLength != null) {
       return exceedingLength(failedValue, max);
     }
@@ -263,38 +260,30 @@ class _$ExceedingLength<T> implements ExceedingLength<T> {
 
   @override
   @optionalTypeArgs
-  TResult map<TResult extends Object>({
-    @required TResult exceedingLength(ExceedingLength<T> value),
-    @required TResult empty(Empty<T> value),
-    @required TResult multiline(Multiline<T> value),
-    @required TResult listTooLong(ListTooLong<T> value),
-    @required TResult invalidEmail(InvalidEmail<T> value),
-    @required TResult shortPassword(ShortPassword<T> value),
-    @required TResult invalidValue(InvalidValue<T> value),
+  TResult map<TResult extends Object?>({
+    required TResult Function(ExceedingLength<T> value) exceedingLength,
+    required TResult Function(Empty<T> value) empty,
+    required TResult Function(Multiline<T> value) multiline,
+    required TResult Function(ListTooLong<T> value) listTooLong,
+    required TResult Function(InvalidEmail<T> value) invalidEmail,
+    required TResult Function(ShortPassword<T> value) shortPassword,
+    required TResult Function(InvalidValue<T> value) invalidValue,
   }) {
-    assert(exceedingLength != null);
-    assert(empty != null);
-    assert(multiline != null);
-    assert(listTooLong != null);
-    assert(invalidEmail != null);
-    assert(shortPassword != null);
-    assert(invalidValue != null);
     return exceedingLength(this);
   }
 
   @override
   @optionalTypeArgs
-  TResult maybeMap<TResult extends Object>({
-    TResult exceedingLength(ExceedingLength<T> value),
-    TResult empty(Empty<T> value),
-    TResult multiline(Multiline<T> value),
-    TResult listTooLong(ListTooLong<T> value),
-    TResult invalidEmail(InvalidEmail<T> value),
-    TResult shortPassword(ShortPassword<T> value),
-    TResult invalidValue(InvalidValue<T> value),
-    @required TResult orElse(),
+  TResult maybeMap<TResult extends Object?>({
+    TResult Function(ExceedingLength<T> value)? exceedingLength,
+    TResult Function(Empty<T> value)? empty,
+    TResult Function(Multiline<T> value)? multiline,
+    TResult Function(ListTooLong<T> value)? listTooLong,
+    TResult Function(InvalidEmail<T> value)? invalidEmail,
+    TResult Function(ShortPassword<T> value)? shortPassword,
+    TResult Function(InvalidValue<T> value)? invalidValue,
+    required TResult orElse(),
   }) {
-    assert(orElse != null);
     if (exceedingLength != null) {
       return exceedingLength(this);
     }
@@ -303,15 +292,16 @@ class _$ExceedingLength<T> implements ExceedingLength<T> {
 }
 
 abstract class ExceedingLength<T> implements ValueFailure<T> {
-  const factory ExceedingLength({@required T failedValue, @required int max}) =
+  const factory ExceedingLength({required T failedValue, required int max}) =
       _$ExceedingLength<T>;
 
   @override
-  T get failedValue;
-  int get max;
+  T get failedValue => throw _privateConstructorUsedError;
+  int get max => throw _privateConstructorUsedError;
   @override
   @JsonKey(ignore: true)
-  $ExceedingLengthCopyWith<T, ExceedingLength<T>> get copyWith;
+  $ExceedingLengthCopyWith<T, ExceedingLength<T>> get copyWith =>
+      throw _privateConstructorUsedError;
 }
 
 /// @nodoc
@@ -334,18 +324,20 @@ class _$EmptyCopyWithImpl<T, $Res> extends _$ValueFailureCopyWithImpl<T, $Res>
 
   @override
   $Res call({
-    Object failedValue = freezed,
+    Object? failedValue = freezed,
   }) {
     return _then(Empty<T>(
-      failedValue:
-          failedValue == freezed ? _value.failedValue : failedValue as T,
+      failedValue: failedValue == freezed
+          ? _value.failedValue
+          : failedValue // ignore: cast_nullable_to_non_nullable
+              as T,
     ));
   }
 }
 
 /// @nodoc
 class _$Empty<T> implements Empty<T> {
-  const _$Empty({@required this.failedValue}) : assert(failedValue != null);
+  const _$Empty({required this.failedValue});
 
   @override
   final T failedValue;
@@ -375,38 +367,30 @@ class _$Empty<T> implements Empty<T> {
 
   @override
   @optionalTypeArgs
-  TResult when<TResult extends Object>({
-    @required TResult exceedingLength(T failedValue, int max),
-    @required TResult empty(T failedValue),
-    @required TResult multiline(T failedValue),
-    @required TResult listTooLong(T failedValue, int max),
-    @required TResult invalidEmail(T failedValue),
-    @required TResult shortPassword(T failedValue),
-    @required TResult invalidValue(T failedValue),
+  TResult when<TResult extends Object?>({
+    required TResult Function(T failedValue, int max) exceedingLength,
+    required TResult Function(T failedValue) empty,
+    required TResult Function(T failedValue) multiline,
+    required TResult Function(T failedValue, int max) listTooLong,
+    required TResult Function(T failedValue) invalidEmail,
+    required TResult Function(T failedValue) shortPassword,
+    required TResult Function(T failedValue) invalidValue,
   }) {
-    assert(exceedingLength != null);
-    assert(empty != null);
-    assert(multiline != null);
-    assert(listTooLong != null);
-    assert(invalidEmail != null);
-    assert(shortPassword != null);
-    assert(invalidValue != null);
     return empty(failedValue);
   }
 
   @override
   @optionalTypeArgs
-  TResult maybeWhen<TResult extends Object>({
-    TResult exceedingLength(T failedValue, int max),
-    TResult empty(T failedValue),
-    TResult multiline(T failedValue),
-    TResult listTooLong(T failedValue, int max),
-    TResult invalidEmail(T failedValue),
-    TResult shortPassword(T failedValue),
-    TResult invalidValue(T failedValue),
-    @required TResult orElse(),
+  TResult maybeWhen<TResult extends Object?>({
+    TResult Function(T failedValue, int max)? exceedingLength,
+    TResult Function(T failedValue)? empty,
+    TResult Function(T failedValue)? multiline,
+    TResult Function(T failedValue, int max)? listTooLong,
+    TResult Function(T failedValue)? invalidEmail,
+    TResult Function(T failedValue)? shortPassword,
+    TResult Function(T failedValue)? invalidValue,
+    required TResult orElse(),
   }) {
-    assert(orElse != null);
     if (empty != null) {
       return empty(failedValue);
     }
@@ -415,38 +399,30 @@ class _$Empty<T> implements Empty<T> {
 
   @override
   @optionalTypeArgs
-  TResult map<TResult extends Object>({
-    @required TResult exceedingLength(ExceedingLength<T> value),
-    @required TResult empty(Empty<T> value),
-    @required TResult multiline(Multiline<T> value),
-    @required TResult listTooLong(ListTooLong<T> value),
-    @required TResult invalidEmail(InvalidEmail<T> value),
-    @required TResult shortPassword(ShortPassword<T> value),
-    @required TResult invalidValue(InvalidValue<T> value),
+  TResult map<TResult extends Object?>({
+    required TResult Function(ExceedingLength<T> value) exceedingLength,
+    required TResult Function(Empty<T> value) empty,
+    required TResult Function(Multiline<T> value) multiline,
+    required TResult Function(ListTooLong<T> value) listTooLong,
+    required TResult Function(InvalidEmail<T> value) invalidEmail,
+    required TResult Function(ShortPassword<T> value) shortPassword,
+    required TResult Function(InvalidValue<T> value) invalidValue,
   }) {
-    assert(exceedingLength != null);
-    assert(empty != null);
-    assert(multiline != null);
-    assert(listTooLong != null);
-    assert(invalidEmail != null);
-    assert(shortPassword != null);
-    assert(invalidValue != null);
     return empty(this);
   }
 
   @override
   @optionalTypeArgs
-  TResult maybeMap<TResult extends Object>({
-    TResult exceedingLength(ExceedingLength<T> value),
-    TResult empty(Empty<T> value),
-    TResult multiline(Multiline<T> value),
-    TResult listTooLong(ListTooLong<T> value),
-    TResult invalidEmail(InvalidEmail<T> value),
-    TResult shortPassword(ShortPassword<T> value),
-    TResult invalidValue(InvalidValue<T> value),
-    @required TResult orElse(),
+  TResult maybeMap<TResult extends Object?>({
+    TResult Function(ExceedingLength<T> value)? exceedingLength,
+    TResult Function(Empty<T> value)? empty,
+    TResult Function(Multiline<T> value)? multiline,
+    TResult Function(ListTooLong<T> value)? listTooLong,
+    TResult Function(InvalidEmail<T> value)? invalidEmail,
+    TResult Function(ShortPassword<T> value)? shortPassword,
+    TResult Function(InvalidValue<T> value)? invalidValue,
+    required TResult orElse(),
   }) {
-    assert(orElse != null);
     if (empty != null) {
       return empty(this);
     }
@@ -455,13 +431,14 @@ class _$Empty<T> implements Empty<T> {
 }
 
 abstract class Empty<T> implements ValueFailure<T> {
-  const factory Empty({@required T failedValue}) = _$Empty<T>;
+  const factory Empty({required T failedValue}) = _$Empty<T>;
 
   @override
-  T get failedValue;
+  T get failedValue => throw _privateConstructorUsedError;
   @override
   @JsonKey(ignore: true)
-  $EmptyCopyWith<T, Empty<T>> get copyWith;
+  $EmptyCopyWith<T, Empty<T>> get copyWith =>
+      throw _privateConstructorUsedError;
 }
 
 /// @nodoc
@@ -487,18 +464,20 @@ class _$MultilineCopyWithImpl<T, $Res>
 
   @override
   $Res call({
-    Object failedValue = freezed,
+    Object? failedValue = freezed,
   }) {
     return _then(Multiline<T>(
-      failedValue:
-          failedValue == freezed ? _value.failedValue : failedValue as T,
+      failedValue: failedValue == freezed
+          ? _value.failedValue
+          : failedValue // ignore: cast_nullable_to_non_nullable
+              as T,
     ));
   }
 }
 
 /// @nodoc
 class _$Multiline<T> implements Multiline<T> {
-  const _$Multiline({@required this.failedValue}) : assert(failedValue != null);
+  const _$Multiline({required this.failedValue});
 
   @override
   final T failedValue;
@@ -528,38 +507,30 @@ class _$Multiline<T> implements Multiline<T> {
 
   @override
   @optionalTypeArgs
-  TResult when<TResult extends Object>({
-    @required TResult exceedingLength(T failedValue, int max),
-    @required TResult empty(T failedValue),
-    @required TResult multiline(T failedValue),
-    @required TResult listTooLong(T failedValue, int max),
-    @required TResult invalidEmail(T failedValue),
-    @required TResult shortPassword(T failedValue),
-    @required TResult invalidValue(T failedValue),
+  TResult when<TResult extends Object?>({
+    required TResult Function(T failedValue, int max) exceedingLength,
+    required TResult Function(T failedValue) empty,
+    required TResult Function(T failedValue) multiline,
+    required TResult Function(T failedValue, int max) listTooLong,
+    required TResult Function(T failedValue) invalidEmail,
+    required TResult Function(T failedValue) shortPassword,
+    required TResult Function(T failedValue) invalidValue,
   }) {
-    assert(exceedingLength != null);
-    assert(empty != null);
-    assert(multiline != null);
-    assert(listTooLong != null);
-    assert(invalidEmail != null);
-    assert(shortPassword != null);
-    assert(invalidValue != null);
     return multiline(failedValue);
   }
 
   @override
   @optionalTypeArgs
-  TResult maybeWhen<TResult extends Object>({
-    TResult exceedingLength(T failedValue, int max),
-    TResult empty(T failedValue),
-    TResult multiline(T failedValue),
-    TResult listTooLong(T failedValue, int max),
-    TResult invalidEmail(T failedValue),
-    TResult shortPassword(T failedValue),
-    TResult invalidValue(T failedValue),
-    @required TResult orElse(),
+  TResult maybeWhen<TResult extends Object?>({
+    TResult Function(T failedValue, int max)? exceedingLength,
+    TResult Function(T failedValue)? empty,
+    TResult Function(T failedValue)? multiline,
+    TResult Function(T failedValue, int max)? listTooLong,
+    TResult Function(T failedValue)? invalidEmail,
+    TResult Function(T failedValue)? shortPassword,
+    TResult Function(T failedValue)? invalidValue,
+    required TResult orElse(),
   }) {
-    assert(orElse != null);
     if (multiline != null) {
       return multiline(failedValue);
     }
@@ -568,38 +539,30 @@ class _$Multiline<T> implements Multiline<T> {
 
   @override
   @optionalTypeArgs
-  TResult map<TResult extends Object>({
-    @required TResult exceedingLength(ExceedingLength<T> value),
-    @required TResult empty(Empty<T> value),
-    @required TResult multiline(Multiline<T> value),
-    @required TResult listTooLong(ListTooLong<T> value),
-    @required TResult invalidEmail(InvalidEmail<T> value),
-    @required TResult shortPassword(ShortPassword<T> value),
-    @required TResult invalidValue(InvalidValue<T> value),
+  TResult map<TResult extends Object?>({
+    required TResult Function(ExceedingLength<T> value) exceedingLength,
+    required TResult Function(Empty<T> value) empty,
+    required TResult Function(Multiline<T> value) multiline,
+    required TResult Function(ListTooLong<T> value) listTooLong,
+    required TResult Function(InvalidEmail<T> value) invalidEmail,
+    required TResult Function(ShortPassword<T> value) shortPassword,
+    required TResult Function(InvalidValue<T> value) invalidValue,
   }) {
-    assert(exceedingLength != null);
-    assert(empty != null);
-    assert(multiline != null);
-    assert(listTooLong != null);
-    assert(invalidEmail != null);
-    assert(shortPassword != null);
-    assert(invalidValue != null);
     return multiline(this);
   }
 
   @override
   @optionalTypeArgs
-  TResult maybeMap<TResult extends Object>({
-    TResult exceedingLength(ExceedingLength<T> value),
-    TResult empty(Empty<T> value),
-    TResult multiline(Multiline<T> value),
-    TResult listTooLong(ListTooLong<T> value),
-    TResult invalidEmail(InvalidEmail<T> value),
-    TResult shortPassword(ShortPassword<T> value),
-    TResult invalidValue(InvalidValue<T> value),
-    @required TResult orElse(),
+  TResult maybeMap<TResult extends Object?>({
+    TResult Function(ExceedingLength<T> value)? exceedingLength,
+    TResult Function(Empty<T> value)? empty,
+    TResult Function(Multiline<T> value)? multiline,
+    TResult Function(ListTooLong<T> value)? listTooLong,
+    TResult Function(InvalidEmail<T> value)? invalidEmail,
+    TResult Function(ShortPassword<T> value)? shortPassword,
+    TResult Function(InvalidValue<T> value)? invalidValue,
+    required TResult orElse(),
   }) {
-    assert(orElse != null);
     if (multiline != null) {
       return multiline(this);
     }
@@ -608,13 +571,14 @@ class _$Multiline<T> implements Multiline<T> {
 }
 
 abstract class Multiline<T> implements ValueFailure<T> {
-  const factory Multiline({@required T failedValue}) = _$Multiline<T>;
+  const factory Multiline({required T failedValue}) = _$Multiline<T>;
 
   @override
-  T get failedValue;
+  T get failedValue => throw _privateConstructorUsedError;
   @override
   @JsonKey(ignore: true)
-  $MultilineCopyWith<T, Multiline<T>> get copyWith;
+  $MultilineCopyWith<T, Multiline<T>> get copyWith =>
+      throw _privateConstructorUsedError;
 }
 
 /// @nodoc
@@ -640,22 +604,25 @@ class _$ListTooLongCopyWithImpl<T, $Res>
 
   @override
   $Res call({
-    Object failedValue = freezed,
-    Object max = freezed,
+    Object? failedValue = freezed,
+    Object? max = freezed,
   }) {
     return _then(ListTooLong<T>(
-      failedValue:
-          failedValue == freezed ? _value.failedValue : failedValue as T,
-      max: max == freezed ? _value.max : max as int,
+      failedValue: failedValue == freezed
+          ? _value.failedValue
+          : failedValue // ignore: cast_nullable_to_non_nullable
+              as T,
+      max: max == freezed
+          ? _value.max
+          : max // ignore: cast_nullable_to_non_nullable
+              as int,
     ));
   }
 }
 
 /// @nodoc
 class _$ListTooLong<T> implements ListTooLong<T> {
-  const _$ListTooLong({@required this.failedValue, @required this.max})
-      : assert(failedValue != null),
-        assert(max != null);
+  const _$ListTooLong({required this.failedValue, required this.max});
 
   @override
   final T failedValue;
@@ -691,38 +658,30 @@ class _$ListTooLong<T> implements ListTooLong<T> {
 
   @override
   @optionalTypeArgs
-  TResult when<TResult extends Object>({
-    @required TResult exceedingLength(T failedValue, int max),
-    @required TResult empty(T failedValue),
-    @required TResult multiline(T failedValue),
-    @required TResult listTooLong(T failedValue, int max),
-    @required TResult invalidEmail(T failedValue),
-    @required TResult shortPassword(T failedValue),
-    @required TResult invalidValue(T failedValue),
+  TResult when<TResult extends Object?>({
+    required TResult Function(T failedValue, int max) exceedingLength,
+    required TResult Function(T failedValue) empty,
+    required TResult Function(T failedValue) multiline,
+    required TResult Function(T failedValue, int max) listTooLong,
+    required TResult Function(T failedValue) invalidEmail,
+    required TResult Function(T failedValue) shortPassword,
+    required TResult Function(T failedValue) invalidValue,
   }) {
-    assert(exceedingLength != null);
-    assert(empty != null);
-    assert(multiline != null);
-    assert(listTooLong != null);
-    assert(invalidEmail != null);
-    assert(shortPassword != null);
-    assert(invalidValue != null);
     return listTooLong(failedValue, max);
   }
 
   @override
   @optionalTypeArgs
-  TResult maybeWhen<TResult extends Object>({
-    TResult exceedingLength(T failedValue, int max),
-    TResult empty(T failedValue),
-    TResult multiline(T failedValue),
-    TResult listTooLong(T failedValue, int max),
-    TResult invalidEmail(T failedValue),
-    TResult shortPassword(T failedValue),
-    TResult invalidValue(T failedValue),
-    @required TResult orElse(),
+  TResult maybeWhen<TResult extends Object?>({
+    TResult Function(T failedValue, int max)? exceedingLength,
+    TResult Function(T failedValue)? empty,
+    TResult Function(T failedValue)? multiline,
+    TResult Function(T failedValue, int max)? listTooLong,
+    TResult Function(T failedValue)? invalidEmail,
+    TResult Function(T failedValue)? shortPassword,
+    TResult Function(T failedValue)? invalidValue,
+    required TResult orElse(),
   }) {
-    assert(orElse != null);
     if (listTooLong != null) {
       return listTooLong(failedValue, max);
     }
@@ -731,38 +690,30 @@ class _$ListTooLong<T> implements ListTooLong<T> {
 
   @override
   @optionalTypeArgs
-  TResult map<TResult extends Object>({
-    @required TResult exceedingLength(ExceedingLength<T> value),
-    @required TResult empty(Empty<T> value),
-    @required TResult multiline(Multiline<T> value),
-    @required TResult listTooLong(ListTooLong<T> value),
-    @required TResult invalidEmail(InvalidEmail<T> value),
-    @required TResult shortPassword(ShortPassword<T> value),
-    @required TResult invalidValue(InvalidValue<T> value),
+  TResult map<TResult extends Object?>({
+    required TResult Function(ExceedingLength<T> value) exceedingLength,
+    required TResult Function(Empty<T> value) empty,
+    required TResult Function(Multiline<T> value) multiline,
+    required TResult Function(ListTooLong<T> value) listTooLong,
+    required TResult Function(InvalidEmail<T> value) invalidEmail,
+    required TResult Function(ShortPassword<T> value) shortPassword,
+    required TResult Function(InvalidValue<T> value) invalidValue,
   }) {
-    assert(exceedingLength != null);
-    assert(empty != null);
-    assert(multiline != null);
-    assert(listTooLong != null);
-    assert(invalidEmail != null);
-    assert(shortPassword != null);
-    assert(invalidValue != null);
     return listTooLong(this);
   }
 
   @override
   @optionalTypeArgs
-  TResult maybeMap<TResult extends Object>({
-    TResult exceedingLength(ExceedingLength<T> value),
-    TResult empty(Empty<T> value),
-    TResult multiline(Multiline<T> value),
-    TResult listTooLong(ListTooLong<T> value),
-    TResult invalidEmail(InvalidEmail<T> value),
-    TResult shortPassword(ShortPassword<T> value),
-    TResult invalidValue(InvalidValue<T> value),
-    @required TResult orElse(),
+  TResult maybeMap<TResult extends Object?>({
+    TResult Function(ExceedingLength<T> value)? exceedingLength,
+    TResult Function(Empty<T> value)? empty,
+    TResult Function(Multiline<T> value)? multiline,
+    TResult Function(ListTooLong<T> value)? listTooLong,
+    TResult Function(InvalidEmail<T> value)? invalidEmail,
+    TResult Function(ShortPassword<T> value)? shortPassword,
+    TResult Function(InvalidValue<T> value)? invalidValue,
+    required TResult orElse(),
   }) {
-    assert(orElse != null);
     if (listTooLong != null) {
       return listTooLong(this);
     }
@@ -771,15 +722,16 @@ class _$ListTooLong<T> implements ListTooLong<T> {
 }
 
 abstract class ListTooLong<T> implements ValueFailure<T> {
-  const factory ListTooLong({@required T failedValue, @required int max}) =
+  const factory ListTooLong({required T failedValue, required int max}) =
       _$ListTooLong<T>;
 
   @override
-  T get failedValue;
-  int get max;
+  T get failedValue => throw _privateConstructorUsedError;
+  int get max => throw _privateConstructorUsedError;
   @override
   @JsonKey(ignore: true)
-  $ListTooLongCopyWith<T, ListTooLong<T>> get copyWith;
+  $ListTooLongCopyWith<T, ListTooLong<T>> get copyWith =>
+      throw _privateConstructorUsedError;
 }
 
 /// @nodoc
@@ -805,19 +757,20 @@ class _$InvalidEmailCopyWithImpl<T, $Res>
 
   @override
   $Res call({
-    Object failedValue = freezed,
+    Object? failedValue = freezed,
   }) {
     return _then(InvalidEmail<T>(
-      failedValue:
-          failedValue == freezed ? _value.failedValue : failedValue as T,
+      failedValue: failedValue == freezed
+          ? _value.failedValue
+          : failedValue // ignore: cast_nullable_to_non_nullable
+              as T,
     ));
   }
 }
 
 /// @nodoc
 class _$InvalidEmail<T> implements InvalidEmail<T> {
-  const _$InvalidEmail({@required this.failedValue})
-      : assert(failedValue != null);
+  const _$InvalidEmail({required this.failedValue});
 
   @override
   final T failedValue;
@@ -847,38 +800,30 @@ class _$InvalidEmail<T> implements InvalidEmail<T> {
 
   @override
   @optionalTypeArgs
-  TResult when<TResult extends Object>({
-    @required TResult exceedingLength(T failedValue, int max),
-    @required TResult empty(T failedValue),
-    @required TResult multiline(T failedValue),
-    @required TResult listTooLong(T failedValue, int max),
-    @required TResult invalidEmail(T failedValue),
-    @required TResult shortPassword(T failedValue),
-    @required TResult invalidValue(T failedValue),
+  TResult when<TResult extends Object?>({
+    required TResult Function(T failedValue, int max) exceedingLength,
+    required TResult Function(T failedValue) empty,
+    required TResult Function(T failedValue) multiline,
+    required TResult Function(T failedValue, int max) listTooLong,
+    required TResult Function(T failedValue) invalidEmail,
+    required TResult Function(T failedValue) shortPassword,
+    required TResult Function(T failedValue) invalidValue,
   }) {
-    assert(exceedingLength != null);
-    assert(empty != null);
-    assert(multiline != null);
-    assert(listTooLong != null);
-    assert(invalidEmail != null);
-    assert(shortPassword != null);
-    assert(invalidValue != null);
     return invalidEmail(failedValue);
   }
 
   @override
   @optionalTypeArgs
-  TResult maybeWhen<TResult extends Object>({
-    TResult exceedingLength(T failedValue, int max),
-    TResult empty(T failedValue),
-    TResult multiline(T failedValue),
-    TResult listTooLong(T failedValue, int max),
-    TResult invalidEmail(T failedValue),
-    TResult shortPassword(T failedValue),
-    TResult invalidValue(T failedValue),
-    @required TResult orElse(),
+  TResult maybeWhen<TResult extends Object?>({
+    TResult Function(T failedValue, int max)? exceedingLength,
+    TResult Function(T failedValue)? empty,
+    TResult Function(T failedValue)? multiline,
+    TResult Function(T failedValue, int max)? listTooLong,
+    TResult Function(T failedValue)? invalidEmail,
+    TResult Function(T failedValue)? shortPassword,
+    TResult Function(T failedValue)? invalidValue,
+    required TResult orElse(),
   }) {
-    assert(orElse != null);
     if (invalidEmail != null) {
       return invalidEmail(failedValue);
     }
@@ -887,38 +832,30 @@ class _$InvalidEmail<T> implements InvalidEmail<T> {
 
   @override
   @optionalTypeArgs
-  TResult map<TResult extends Object>({
-    @required TResult exceedingLength(ExceedingLength<T> value),
-    @required TResult empty(Empty<T> value),
-    @required TResult multiline(Multiline<T> value),
-    @required TResult listTooLong(ListTooLong<T> value),
-    @required TResult invalidEmail(InvalidEmail<T> value),
-    @required TResult shortPassword(ShortPassword<T> value),
-    @required TResult invalidValue(InvalidValue<T> value),
+  TResult map<TResult extends Object?>({
+    required TResult Function(ExceedingLength<T> value) exceedingLength,
+    required TResult Function(Empty<T> value) empty,
+    required TResult Function(Multiline<T> value) multiline,
+    required TResult Function(ListTooLong<T> value) listTooLong,
+    required TResult Function(InvalidEmail<T> value) invalidEmail,
+    required TResult Function(ShortPassword<T> value) shortPassword,
+    required TResult Function(InvalidValue<T> value) invalidValue,
   }) {
-    assert(exceedingLength != null);
-    assert(empty != null);
-    assert(multiline != null);
-    assert(listTooLong != null);
-    assert(invalidEmail != null);
-    assert(shortPassword != null);
-    assert(invalidValue != null);
     return invalidEmail(this);
   }
 
   @override
   @optionalTypeArgs
-  TResult maybeMap<TResult extends Object>({
-    TResult exceedingLength(ExceedingLength<T> value),
-    TResult empty(Empty<T> value),
-    TResult multiline(Multiline<T> value),
-    TResult listTooLong(ListTooLong<T> value),
-    TResult invalidEmail(InvalidEmail<T> value),
-    TResult shortPassword(ShortPassword<T> value),
-    TResult invalidValue(InvalidValue<T> value),
-    @required TResult orElse(),
+  TResult maybeMap<TResult extends Object?>({
+    TResult Function(ExceedingLength<T> value)? exceedingLength,
+    TResult Function(Empty<T> value)? empty,
+    TResult Function(Multiline<T> value)? multiline,
+    TResult Function(ListTooLong<T> value)? listTooLong,
+    TResult Function(InvalidEmail<T> value)? invalidEmail,
+    TResult Function(ShortPassword<T> value)? shortPassword,
+    TResult Function(InvalidValue<T> value)? invalidValue,
+    required TResult orElse(),
   }) {
-    assert(orElse != null);
     if (invalidEmail != null) {
       return invalidEmail(this);
     }
@@ -927,13 +864,14 @@ class _$InvalidEmail<T> implements InvalidEmail<T> {
 }
 
 abstract class InvalidEmail<T> implements ValueFailure<T> {
-  const factory InvalidEmail({@required T failedValue}) = _$InvalidEmail<T>;
+  const factory InvalidEmail({required T failedValue}) = _$InvalidEmail<T>;
 
   @override
-  T get failedValue;
+  T get failedValue => throw _privateConstructorUsedError;
   @override
   @JsonKey(ignore: true)
-  $InvalidEmailCopyWith<T, InvalidEmail<T>> get copyWith;
+  $InvalidEmailCopyWith<T, InvalidEmail<T>> get copyWith =>
+      throw _privateConstructorUsedError;
 }
 
 /// @nodoc
@@ -959,19 +897,20 @@ class _$ShortPasswordCopyWithImpl<T, $Res>
 
   @override
   $Res call({
-    Object failedValue = freezed,
+    Object? failedValue = freezed,
   }) {
     return _then(ShortPassword<T>(
-      failedValue:
-          failedValue == freezed ? _value.failedValue : failedValue as T,
+      failedValue: failedValue == freezed
+          ? _value.failedValue
+          : failedValue // ignore: cast_nullable_to_non_nullable
+              as T,
     ));
   }
 }
 
 /// @nodoc
 class _$ShortPassword<T> implements ShortPassword<T> {
-  const _$ShortPassword({@required this.failedValue})
-      : assert(failedValue != null);
+  const _$ShortPassword({required this.failedValue});
 
   @override
   final T failedValue;
@@ -1001,38 +940,30 @@ class _$ShortPassword<T> implements ShortPassword<T> {
 
   @override
   @optionalTypeArgs
-  TResult when<TResult extends Object>({
-    @required TResult exceedingLength(T failedValue, int max),
-    @required TResult empty(T failedValue),
-    @required TResult multiline(T failedValue),
-    @required TResult listTooLong(T failedValue, int max),
-    @required TResult invalidEmail(T failedValue),
-    @required TResult shortPassword(T failedValue),
-    @required TResult invalidValue(T failedValue),
+  TResult when<TResult extends Object?>({
+    required TResult Function(T failedValue, int max) exceedingLength,
+    required TResult Function(T failedValue) empty,
+    required TResult Function(T failedValue) multiline,
+    required TResult Function(T failedValue, int max) listTooLong,
+    required TResult Function(T failedValue) invalidEmail,
+    required TResult Function(T failedValue) shortPassword,
+    required TResult Function(T failedValue) invalidValue,
   }) {
-    assert(exceedingLength != null);
-    assert(empty != null);
-    assert(multiline != null);
-    assert(listTooLong != null);
-    assert(invalidEmail != null);
-    assert(shortPassword != null);
-    assert(invalidValue != null);
     return shortPassword(failedValue);
   }
 
   @override
   @optionalTypeArgs
-  TResult maybeWhen<TResult extends Object>({
-    TResult exceedingLength(T failedValue, int max),
-    TResult empty(T failedValue),
-    TResult multiline(T failedValue),
-    TResult listTooLong(T failedValue, int max),
-    TResult invalidEmail(T failedValue),
-    TResult shortPassword(T failedValue),
-    TResult invalidValue(T failedValue),
-    @required TResult orElse(),
+  TResult maybeWhen<TResult extends Object?>({
+    TResult Function(T failedValue, int max)? exceedingLength,
+    TResult Function(T failedValue)? empty,
+    TResult Function(T failedValue)? multiline,
+    TResult Function(T failedValue, int max)? listTooLong,
+    TResult Function(T failedValue)? invalidEmail,
+    TResult Function(T failedValue)? shortPassword,
+    TResult Function(T failedValue)? invalidValue,
+    required TResult orElse(),
   }) {
-    assert(orElse != null);
     if (shortPassword != null) {
       return shortPassword(failedValue);
     }
@@ -1041,38 +972,30 @@ class _$ShortPassword<T> implements ShortPassword<T> {
 
   @override
   @optionalTypeArgs
-  TResult map<TResult extends Object>({
-    @required TResult exceedingLength(ExceedingLength<T> value),
-    @required TResult empty(Empty<T> value),
-    @required TResult multiline(Multiline<T> value),
-    @required TResult listTooLong(ListTooLong<T> value),
-    @required TResult invalidEmail(InvalidEmail<T> value),
-    @required TResult shortPassword(ShortPassword<T> value),
-    @required TResult invalidValue(InvalidValue<T> value),
+  TResult map<TResult extends Object?>({
+    required TResult Function(ExceedingLength<T> value) exceedingLength,
+    required TResult Function(Empty<T> value) empty,
+    required TResult Function(Multiline<T> value) multiline,
+    required TResult Function(ListTooLong<T> value) listTooLong,
+    required TResult Function(InvalidEmail<T> value) invalidEmail,
+    required TResult Function(ShortPassword<T> value) shortPassword,
+    required TResult Function(InvalidValue<T> value) invalidValue,
   }) {
-    assert(exceedingLength != null);
-    assert(empty != null);
-    assert(multiline != null);
-    assert(listTooLong != null);
-    assert(invalidEmail != null);
-    assert(shortPassword != null);
-    assert(invalidValue != null);
     return shortPassword(this);
   }
 
   @override
   @optionalTypeArgs
-  TResult maybeMap<TResult extends Object>({
-    TResult exceedingLength(ExceedingLength<T> value),
-    TResult empty(Empty<T> value),
-    TResult multiline(Multiline<T> value),
-    TResult listTooLong(ListTooLong<T> value),
-    TResult invalidEmail(InvalidEmail<T> value),
-    TResult shortPassword(ShortPassword<T> value),
-    TResult invalidValue(InvalidValue<T> value),
-    @required TResult orElse(),
+  TResult maybeMap<TResult extends Object?>({
+    TResult Function(ExceedingLength<T> value)? exceedingLength,
+    TResult Function(Empty<T> value)? empty,
+    TResult Function(Multiline<T> value)? multiline,
+    TResult Function(ListTooLong<T> value)? listTooLong,
+    TResult Function(InvalidEmail<T> value)? invalidEmail,
+    TResult Function(ShortPassword<T> value)? shortPassword,
+    TResult Function(InvalidValue<T> value)? invalidValue,
+    required TResult orElse(),
   }) {
-    assert(orElse != null);
     if (shortPassword != null) {
       return shortPassword(this);
     }
@@ -1081,13 +1004,14 @@ class _$ShortPassword<T> implements ShortPassword<T> {
 }
 
 abstract class ShortPassword<T> implements ValueFailure<T> {
-  const factory ShortPassword({@required T failedValue}) = _$ShortPassword<T>;
+  const factory ShortPassword({required T failedValue}) = _$ShortPassword<T>;
 
   @override
-  T get failedValue;
+  T get failedValue => throw _privateConstructorUsedError;
   @override
   @JsonKey(ignore: true)
-  $ShortPasswordCopyWith<T, ShortPassword<T>> get copyWith;
+  $ShortPasswordCopyWith<T, ShortPassword<T>> get copyWith =>
+      throw _privateConstructorUsedError;
 }
 
 /// @nodoc
@@ -1113,19 +1037,20 @@ class _$InvalidValueCopyWithImpl<T, $Res>
 
   @override
   $Res call({
-    Object failedValue = freezed,
+    Object? failedValue = freezed,
   }) {
     return _then(InvalidValue<T>(
-      failedValue:
-          failedValue == freezed ? _value.failedValue : failedValue as T,
+      failedValue: failedValue == freezed
+          ? _value.failedValue
+          : failedValue // ignore: cast_nullable_to_non_nullable
+              as T,
     ));
   }
 }
 
 /// @nodoc
 class _$InvalidValue<T> implements InvalidValue<T> {
-  const _$InvalidValue({@required this.failedValue})
-      : assert(failedValue != null);
+  const _$InvalidValue({required this.failedValue});
 
   @override
   final T failedValue;
@@ -1155,38 +1080,30 @@ class _$InvalidValue<T> implements InvalidValue<T> {
 
   @override
   @optionalTypeArgs
-  TResult when<TResult extends Object>({
-    @required TResult exceedingLength(T failedValue, int max),
-    @required TResult empty(T failedValue),
-    @required TResult multiline(T failedValue),
-    @required TResult listTooLong(T failedValue, int max),
-    @required TResult invalidEmail(T failedValue),
-    @required TResult shortPassword(T failedValue),
-    @required TResult invalidValue(T failedValue),
+  TResult when<TResult extends Object?>({
+    required TResult Function(T failedValue, int max) exceedingLength,
+    required TResult Function(T failedValue) empty,
+    required TResult Function(T failedValue) multiline,
+    required TResult Function(T failedValue, int max) listTooLong,
+    required TResult Function(T failedValue) invalidEmail,
+    required TResult Function(T failedValue) shortPassword,
+    required TResult Function(T failedValue) invalidValue,
   }) {
-    assert(exceedingLength != null);
-    assert(empty != null);
-    assert(multiline != null);
-    assert(listTooLong != null);
-    assert(invalidEmail != null);
-    assert(shortPassword != null);
-    assert(invalidValue != null);
     return invalidValue(failedValue);
   }
 
   @override
   @optionalTypeArgs
-  TResult maybeWhen<TResult extends Object>({
-    TResult exceedingLength(T failedValue, int max),
-    TResult empty(T failedValue),
-    TResult multiline(T failedValue),
-    TResult listTooLong(T failedValue, int max),
-    TResult invalidEmail(T failedValue),
-    TResult shortPassword(T failedValue),
-    TResult invalidValue(T failedValue),
-    @required TResult orElse(),
+  TResult maybeWhen<TResult extends Object?>({
+    TResult Function(T failedValue, int max)? exceedingLength,
+    TResult Function(T failedValue)? empty,
+    TResult Function(T failedValue)? multiline,
+    TResult Function(T failedValue, int max)? listTooLong,
+    TResult Function(T failedValue)? invalidEmail,
+    TResult Function(T failedValue)? shortPassword,
+    TResult Function(T failedValue)? invalidValue,
+    required TResult orElse(),
   }) {
-    assert(orElse != null);
     if (invalidValue != null) {
       return invalidValue(failedValue);
     }
@@ -1195,38 +1112,30 @@ class _$InvalidValue<T> implements InvalidValue<T> {
 
   @override
   @optionalTypeArgs
-  TResult map<TResult extends Object>({
-    @required TResult exceedingLength(ExceedingLength<T> value),
-    @required TResult empty(Empty<T> value),
-    @required TResult multiline(Multiline<T> value),
-    @required TResult listTooLong(ListTooLong<T> value),
-    @required TResult invalidEmail(InvalidEmail<T> value),
-    @required TResult shortPassword(ShortPassword<T> value),
-    @required TResult invalidValue(InvalidValue<T> value),
+  TResult map<TResult extends Object?>({
+    required TResult Function(ExceedingLength<T> value) exceedingLength,
+    required TResult Function(Empty<T> value) empty,
+    required TResult Function(Multiline<T> value) multiline,
+    required TResult Function(ListTooLong<T> value) listTooLong,
+    required TResult Function(InvalidEmail<T> value) invalidEmail,
+    required TResult Function(ShortPassword<T> value) shortPassword,
+    required TResult Function(InvalidValue<T> value) invalidValue,
   }) {
-    assert(exceedingLength != null);
-    assert(empty != null);
-    assert(multiline != null);
-    assert(listTooLong != null);
-    assert(invalidEmail != null);
-    assert(shortPassword != null);
-    assert(invalidValue != null);
     return invalidValue(this);
   }
 
   @override
   @optionalTypeArgs
-  TResult maybeMap<TResult extends Object>({
-    TResult exceedingLength(ExceedingLength<T> value),
-    TResult empty(Empty<T> value),
-    TResult multiline(Multiline<T> value),
-    TResult listTooLong(ListTooLong<T> value),
-    TResult invalidEmail(InvalidEmail<T> value),
-    TResult shortPassword(ShortPassword<T> value),
-    TResult invalidValue(InvalidValue<T> value),
-    @required TResult orElse(),
+  TResult maybeMap<TResult extends Object?>({
+    TResult Function(ExceedingLength<T> value)? exceedingLength,
+    TResult Function(Empty<T> value)? empty,
+    TResult Function(Multiline<T> value)? multiline,
+    TResult Function(ListTooLong<T> value)? listTooLong,
+    TResult Function(InvalidEmail<T> value)? invalidEmail,
+    TResult Function(ShortPassword<T> value)? shortPassword,
+    TResult Function(InvalidValue<T> value)? invalidValue,
+    required TResult orElse(),
   }) {
-    assert(orElse != null);
     if (invalidValue != null) {
       return invalidValue(this);
     }
@@ -1235,11 +1144,12 @@ class _$InvalidValue<T> implements InvalidValue<T> {
 }
 
 abstract class InvalidValue<T> implements ValueFailure<T> {
-  const factory InvalidValue({@required T failedValue}) = _$InvalidValue<T>;
+  const factory InvalidValue({required T failedValue}) = _$InvalidValue<T>;
 
   @override
-  T get failedValue;
+  T get failedValue => throw _privateConstructorUsedError;
   @override
   @JsonKey(ignore: true)
-  $InvalidValueCopyWith<T, InvalidValue<T>> get copyWith;
+  $InvalidValueCopyWith<T, InvalidValue<T>> get copyWith =>
+      throw _privateConstructorUsedError;
 }
