@@ -1,5 +1,5 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
-// ignore_for_file: deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies
+// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides
 
 part of 'account_dto.dart';
 
@@ -8,6 +8,10 @@ part of 'account_dto.dart';
 // **************************************************************************
 
 T _$identity<T>(T value) => value;
+
+final _privateConstructorUsedError = UnsupportedError(
+    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more informations: https://github.com/rrousselGit/freezed#custom-getters-and-methods');
+
 AccountDto _$AccountDtoFromJson(Map<String, dynamic> json) {
   return _AccountDto.fromJson(json);
 }
@@ -16,24 +20,23 @@ AccountDto _$AccountDtoFromJson(Map<String, dynamic> json) {
 class _$AccountDtoTearOff {
   const _$AccountDtoTearOff();
 
-// ignore: unused_element
   _AccountDto call(
-      {int id,
-      String name,
-      String lastName,
-      int typeUser,
-      String address,
-      String picture,
-      String phone,
-      String idNumber,
-      String userName,
-      String email,
-      String parent,
-      String parentPhone,
-      String experience,
-      DateTime emailVerifiedAt,
-      DateTime createdAt,
-      DateTime updatedAt}) {
+      {int? id,
+      String? name,
+      String? lastName,
+      int? typeUser,
+      String? address,
+      String? picture,
+      String? phone,
+      String? idNumber,
+      String? userName,
+      String? email,
+      String? parent,
+      String? parentPhone,
+      String? experience,
+      DateTime? emailVerifiedAt,
+      DateTime? createdAt,
+      DateTime? updatedAt}) {
     return _AccountDto(
       id: id,
       name: name,
@@ -54,38 +57,37 @@ class _$AccountDtoTearOff {
     );
   }
 
-// ignore: unused_element
   AccountDto fromJson(Map<String, Object> json) {
     return AccountDto.fromJson(json);
   }
 }
 
 /// @nodoc
-// ignore: unused_element
 const $AccountDto = _$AccountDtoTearOff();
 
 /// @nodoc
 mixin _$AccountDto {
-  int get id;
-  String get name;
-  String get lastName;
-  int get typeUser;
-  String get address;
-  String get picture;
-  String get phone;
-  String get idNumber;
-  String get userName;
-  String get email;
-  String get parent;
-  String get parentPhone;
-  String get experience;
-  DateTime get emailVerifiedAt;
-  DateTime get createdAt;
-  DateTime get updatedAt;
+  int? get id => throw _privateConstructorUsedError;
+  String? get name => throw _privateConstructorUsedError;
+  String? get lastName => throw _privateConstructorUsedError;
+  int? get typeUser => throw _privateConstructorUsedError;
+  String? get address => throw _privateConstructorUsedError;
+  String? get picture => throw _privateConstructorUsedError;
+  String? get phone => throw _privateConstructorUsedError;
+  String? get idNumber => throw _privateConstructorUsedError;
+  String? get userName => throw _privateConstructorUsedError;
+  String? get email => throw _privateConstructorUsedError;
+  String? get parent => throw _privateConstructorUsedError;
+  String? get parentPhone => throw _privateConstructorUsedError;
+  String? get experience => throw _privateConstructorUsedError;
+  DateTime? get emailVerifiedAt => throw _privateConstructorUsedError;
+  DateTime? get createdAt => throw _privateConstructorUsedError;
+  DateTime? get updatedAt => throw _privateConstructorUsedError;
 
-  Map<String, dynamic> toJson();
+  Map<String, dynamic> toJson() => throw _privateConstructorUsedError;
   @JsonKey(ignore: true)
-  $AccountDtoCopyWith<AccountDto> get copyWith;
+  $AccountDtoCopyWith<AccountDto> get copyWith =>
+      throw _privateConstructorUsedError;
 }
 
 /// @nodoc
@@ -94,22 +96,22 @@ abstract class $AccountDtoCopyWith<$Res> {
           AccountDto value, $Res Function(AccountDto) then) =
       _$AccountDtoCopyWithImpl<$Res>;
   $Res call(
-      {int id,
-      String name,
-      String lastName,
-      int typeUser,
-      String address,
-      String picture,
-      String phone,
-      String idNumber,
-      String userName,
-      String email,
-      String parent,
-      String parentPhone,
-      String experience,
-      DateTime emailVerifiedAt,
-      DateTime createdAt,
-      DateTime updatedAt});
+      {int? id,
+      String? name,
+      String? lastName,
+      int? typeUser,
+      String? address,
+      String? picture,
+      String? phone,
+      String? idNumber,
+      String? userName,
+      String? email,
+      String? parent,
+      String? parentPhone,
+      String? experience,
+      DateTime? emailVerifiedAt,
+      DateTime? createdAt,
+      DateTime? updatedAt});
 }
 
 /// @nodoc
@@ -122,46 +124,88 @@ class _$AccountDtoCopyWithImpl<$Res> implements $AccountDtoCopyWith<$Res> {
 
   @override
   $Res call({
-    Object id = freezed,
-    Object name = freezed,
-    Object lastName = freezed,
-    Object typeUser = freezed,
-    Object address = freezed,
-    Object picture = freezed,
-    Object phone = freezed,
-    Object idNumber = freezed,
-    Object userName = freezed,
-    Object email = freezed,
-    Object parent = freezed,
-    Object parentPhone = freezed,
-    Object experience = freezed,
-    Object emailVerifiedAt = freezed,
-    Object createdAt = freezed,
-    Object updatedAt = freezed,
+    Object? id = freezed,
+    Object? name = freezed,
+    Object? lastName = freezed,
+    Object? typeUser = freezed,
+    Object? address = freezed,
+    Object? picture = freezed,
+    Object? phone = freezed,
+    Object? idNumber = freezed,
+    Object? userName = freezed,
+    Object? email = freezed,
+    Object? parent = freezed,
+    Object? parentPhone = freezed,
+    Object? experience = freezed,
+    Object? emailVerifiedAt = freezed,
+    Object? createdAt = freezed,
+    Object? updatedAt = freezed,
   }) {
     return _then(_value.copyWith(
-      id: id == freezed ? _value.id : id as int,
-      name: name == freezed ? _value.name : name as String,
-      lastName: lastName == freezed ? _value.lastName : lastName as String,
-      typeUser: typeUser == freezed ? _value.typeUser : typeUser as int,
-      address: address == freezed ? _value.address : address as String,
-      picture: picture == freezed ? _value.picture : picture as String,
-      phone: phone == freezed ? _value.phone : phone as String,
-      idNumber: idNumber == freezed ? _value.idNumber : idNumber as String,
-      userName: userName == freezed ? _value.userName : userName as String,
-      email: email == freezed ? _value.email : email as String,
-      parent: parent == freezed ? _value.parent : parent as String,
-      parentPhone:
-          parentPhone == freezed ? _value.parentPhone : parentPhone as String,
-      experience:
-          experience == freezed ? _value.experience : experience as String,
+      id: id == freezed
+          ? _value.id
+          : id // ignore: cast_nullable_to_non_nullable
+              as int?,
+      name: name == freezed
+          ? _value.name
+          : name // ignore: cast_nullable_to_non_nullable
+              as String?,
+      lastName: lastName == freezed
+          ? _value.lastName
+          : lastName // ignore: cast_nullable_to_non_nullable
+              as String?,
+      typeUser: typeUser == freezed
+          ? _value.typeUser
+          : typeUser // ignore: cast_nullable_to_non_nullable
+              as int?,
+      address: address == freezed
+          ? _value.address
+          : address // ignore: cast_nullable_to_non_nullable
+              as String?,
+      picture: picture == freezed
+          ? _value.picture
+          : picture // ignore: cast_nullable_to_non_nullable
+              as String?,
+      phone: phone == freezed
+          ? _value.phone
+          : phone // ignore: cast_nullable_to_non_nullable
+              as String?,
+      idNumber: idNumber == freezed
+          ? _value.idNumber
+          : idNumber // ignore: cast_nullable_to_non_nullable
+              as String?,
+      userName: userName == freezed
+          ? _value.userName
+          : userName // ignore: cast_nullable_to_non_nullable
+              as String?,
+      email: email == freezed
+          ? _value.email
+          : email // ignore: cast_nullable_to_non_nullable
+              as String?,
+      parent: parent == freezed
+          ? _value.parent
+          : parent // ignore: cast_nullable_to_non_nullable
+              as String?,
+      parentPhone: parentPhone == freezed
+          ? _value.parentPhone
+          : parentPhone // ignore: cast_nullable_to_non_nullable
+              as String?,
+      experience: experience == freezed
+          ? _value.experience
+          : experience // ignore: cast_nullable_to_non_nullable
+              as String?,
       emailVerifiedAt: emailVerifiedAt == freezed
           ? _value.emailVerifiedAt
-          : emailVerifiedAt as DateTime,
-      createdAt:
-          createdAt == freezed ? _value.createdAt : createdAt as DateTime,
-      updatedAt:
-          updatedAt == freezed ? _value.updatedAt : updatedAt as DateTime,
+          : emailVerifiedAt // ignore: cast_nullable_to_non_nullable
+              as DateTime?,
+      createdAt: createdAt == freezed
+          ? _value.createdAt
+          : createdAt // ignore: cast_nullable_to_non_nullable
+              as DateTime?,
+      updatedAt: updatedAt == freezed
+          ? _value.updatedAt
+          : updatedAt // ignore: cast_nullable_to_non_nullable
+              as DateTime?,
     ));
   }
 }
@@ -173,22 +217,22 @@ abstract class _$AccountDtoCopyWith<$Res> implements $AccountDtoCopyWith<$Res> {
       __$AccountDtoCopyWithImpl<$Res>;
   @override
   $Res call(
-      {int id,
-      String name,
-      String lastName,
-      int typeUser,
-      String address,
-      String picture,
-      String phone,
-      String idNumber,
-      String userName,
-      String email,
-      String parent,
-      String parentPhone,
-      String experience,
-      DateTime emailVerifiedAt,
-      DateTime createdAt,
-      DateTime updatedAt});
+      {int? id,
+      String? name,
+      String? lastName,
+      int? typeUser,
+      String? address,
+      String? picture,
+      String? phone,
+      String? idNumber,
+      String? userName,
+      String? email,
+      String? parent,
+      String? parentPhone,
+      String? experience,
+      DateTime? emailVerifiedAt,
+      DateTime? createdAt,
+      DateTime? updatedAt});
 }
 
 /// @nodoc
@@ -203,46 +247,88 @@ class __$AccountDtoCopyWithImpl<$Res> extends _$AccountDtoCopyWithImpl<$Res>
 
   @override
   $Res call({
-    Object id = freezed,
-    Object name = freezed,
-    Object lastName = freezed,
-    Object typeUser = freezed,
-    Object address = freezed,
-    Object picture = freezed,
-    Object phone = freezed,
-    Object idNumber = freezed,
-    Object userName = freezed,
-    Object email = freezed,
-    Object parent = freezed,
-    Object parentPhone = freezed,
-    Object experience = freezed,
-    Object emailVerifiedAt = freezed,
-    Object createdAt = freezed,
-    Object updatedAt = freezed,
+    Object? id = freezed,
+    Object? name = freezed,
+    Object? lastName = freezed,
+    Object? typeUser = freezed,
+    Object? address = freezed,
+    Object? picture = freezed,
+    Object? phone = freezed,
+    Object? idNumber = freezed,
+    Object? userName = freezed,
+    Object? email = freezed,
+    Object? parent = freezed,
+    Object? parentPhone = freezed,
+    Object? experience = freezed,
+    Object? emailVerifiedAt = freezed,
+    Object? createdAt = freezed,
+    Object? updatedAt = freezed,
   }) {
     return _then(_AccountDto(
-      id: id == freezed ? _value.id : id as int,
-      name: name == freezed ? _value.name : name as String,
-      lastName: lastName == freezed ? _value.lastName : lastName as String,
-      typeUser: typeUser == freezed ? _value.typeUser : typeUser as int,
-      address: address == freezed ? _value.address : address as String,
-      picture: picture == freezed ? _value.picture : picture as String,
-      phone: phone == freezed ? _value.phone : phone as String,
-      idNumber: idNumber == freezed ? _value.idNumber : idNumber as String,
-      userName: userName == freezed ? _value.userName : userName as String,
-      email: email == freezed ? _value.email : email as String,
-      parent: parent == freezed ? _value.parent : parent as String,
-      parentPhone:
-          parentPhone == freezed ? _value.parentPhone : parentPhone as String,
-      experience:
-          experience == freezed ? _value.experience : experience as String,
+      id: id == freezed
+          ? _value.id
+          : id // ignore: cast_nullable_to_non_nullable
+              as int?,
+      name: name == freezed
+          ? _value.name
+          : name // ignore: cast_nullable_to_non_nullable
+              as String?,
+      lastName: lastName == freezed
+          ? _value.lastName
+          : lastName // ignore: cast_nullable_to_non_nullable
+              as String?,
+      typeUser: typeUser == freezed
+          ? _value.typeUser
+          : typeUser // ignore: cast_nullable_to_non_nullable
+              as int?,
+      address: address == freezed
+          ? _value.address
+          : address // ignore: cast_nullable_to_non_nullable
+              as String?,
+      picture: picture == freezed
+          ? _value.picture
+          : picture // ignore: cast_nullable_to_non_nullable
+              as String?,
+      phone: phone == freezed
+          ? _value.phone
+          : phone // ignore: cast_nullable_to_non_nullable
+              as String?,
+      idNumber: idNumber == freezed
+          ? _value.idNumber
+          : idNumber // ignore: cast_nullable_to_non_nullable
+              as String?,
+      userName: userName == freezed
+          ? _value.userName
+          : userName // ignore: cast_nullable_to_non_nullable
+              as String?,
+      email: email == freezed
+          ? _value.email
+          : email // ignore: cast_nullable_to_non_nullable
+              as String?,
+      parent: parent == freezed
+          ? _value.parent
+          : parent // ignore: cast_nullable_to_non_nullable
+              as String?,
+      parentPhone: parentPhone == freezed
+          ? _value.parentPhone
+          : parentPhone // ignore: cast_nullable_to_non_nullable
+              as String?,
+      experience: experience == freezed
+          ? _value.experience
+          : experience // ignore: cast_nullable_to_non_nullable
+              as String?,
       emailVerifiedAt: emailVerifiedAt == freezed
           ? _value.emailVerifiedAt
-          : emailVerifiedAt as DateTime,
-      createdAt:
-          createdAt == freezed ? _value.createdAt : createdAt as DateTime,
-      updatedAt:
-          updatedAt == freezed ? _value.updatedAt : updatedAt as DateTime,
+          : emailVerifiedAt // ignore: cast_nullable_to_non_nullable
+              as DateTime?,
+      createdAt: createdAt == freezed
+          ? _value.createdAt
+          : createdAt // ignore: cast_nullable_to_non_nullable
+              as DateTime?,
+      updatedAt: updatedAt == freezed
+          ? _value.updatedAt
+          : updatedAt // ignore: cast_nullable_to_non_nullable
+              as DateTime?,
     ));
   }
 }
@@ -274,37 +360,37 @@ class _$_AccountDto extends _AccountDto {
       _$_$_AccountDtoFromJson(json);
 
   @override
-  final int id;
+  final int? id;
   @override
-  final String name;
+  final String? name;
   @override
-  final String lastName;
+  final String? lastName;
   @override
-  final int typeUser;
+  final int? typeUser;
   @override
-  final String address;
+  final String? address;
   @override
-  final String picture;
+  final String? picture;
   @override
-  final String phone;
+  final String? phone;
   @override
-  final String idNumber;
+  final String? idNumber;
   @override
-  final String userName;
+  final String? userName;
   @override
-  final String email;
+  final String? email;
   @override
-  final String parent;
+  final String? parent;
   @override
-  final String parentPhone;
+  final String? parentPhone;
   @override
-  final String experience;
+  final String? experience;
   @override
-  final DateTime emailVerifiedAt;
+  final DateTime? emailVerifiedAt;
   @override
-  final DateTime createdAt;
+  final DateTime? createdAt;
   @override
-  final DateTime updatedAt;
+  final DateTime? updatedAt;
 
   @override
   String toString() {
@@ -392,61 +478,62 @@ class _$_AccountDto extends _AccountDto {
 }
 
 abstract class _AccountDto extends AccountDto {
-  const _AccountDto._() : super._();
   const factory _AccountDto(
-      {int id,
-      String name,
-      String lastName,
-      int typeUser,
-      String address,
-      String picture,
-      String phone,
-      String idNumber,
-      String userName,
-      String email,
-      String parent,
-      String parentPhone,
-      String experience,
-      DateTime emailVerifiedAt,
-      DateTime createdAt,
-      DateTime updatedAt}) = _$_AccountDto;
+      {int? id,
+      String? name,
+      String? lastName,
+      int? typeUser,
+      String? address,
+      String? picture,
+      String? phone,
+      String? idNumber,
+      String? userName,
+      String? email,
+      String? parent,
+      String? parentPhone,
+      String? experience,
+      DateTime? emailVerifiedAt,
+      DateTime? createdAt,
+      DateTime? updatedAt}) = _$_AccountDto;
+  const _AccountDto._() : super._();
 
   factory _AccountDto.fromJson(Map<String, dynamic> json) =
       _$_AccountDto.fromJson;
 
   @override
-  int get id;
+  int? get id => throw _privateConstructorUsedError;
   @override
-  String get name;
+  String? get name => throw _privateConstructorUsedError;
   @override
-  String get lastName;
+  String? get lastName => throw _privateConstructorUsedError;
   @override
-  int get typeUser;
+  int? get typeUser => throw _privateConstructorUsedError;
   @override
-  String get address;
+  String? get address => throw _privateConstructorUsedError;
   @override
-  String get picture;
+  String? get picture => throw _privateConstructorUsedError;
   @override
-  String get phone;
+  String? get phone => throw _privateConstructorUsedError;
   @override
-  String get idNumber;
+  String? get idNumber => throw _privateConstructorUsedError;
   @override
-  String get userName;
+  String? get userName => throw _privateConstructorUsedError;
   @override
-  String get email;
+  String? get email => throw _privateConstructorUsedError;
   @override
-  String get parent;
+  String? get parent => throw _privateConstructorUsedError;
   @override
-  String get parentPhone;
+  String? get parentPhone => throw _privateConstructorUsedError;
   @override
-  String get experience;
+  String? get experience => throw _privateConstructorUsedError;
   @override
-  DateTime get emailVerifiedAt;
+  DateTime? get emailVerifiedAt => throw _privateConstructorUsedError;
   @override
-  DateTime get createdAt;
+  DateTime? get createdAt => throw _privateConstructorUsedError;
   @override
-  DateTime get updatedAt;
+  DateTime? get updatedAt => throw _privateConstructorUsedError;
   @override
   @JsonKey(ignore: true)
-  _$AccountDtoCopyWith<_AccountDto> get copyWith;
+  _$AccountDtoCopyWith<_AccountDto> get copyWith =>
+      throw _privateConstructorUsedError;
 }
