@@ -1,5 +1,5 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
-// ignore_for_file: deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies
+// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides
 
 part of 'auth_token_dto.dart';
 
@@ -8,6 +8,10 @@ part of 'auth_token_dto.dart';
 // **************************************************************************
 
 T _$identity<T>(T value) => value;
+
+final _privateConstructorUsedError = UnsupportedError(
+    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more informations: https://github.com/rrousselGit/freezed#custom-getters-and-methods');
+
 AuthTokenDto _$AuthTokenDtoFromJson(Map<String, dynamic> json) {
   return _AuthTokenDto.fromJson(json);
 }
@@ -16,30 +20,28 @@ AuthTokenDto _$AuthTokenDtoFromJson(Map<String, dynamic> json) {
 class _$AuthTokenDtoTearOff {
   const _$AuthTokenDtoTearOff();
 
-// ignore: unused_element
-  _AuthTokenDto call({@required String accessToken}) {
+  _AuthTokenDto call({required String accessToken}) {
     return _AuthTokenDto(
       accessToken: accessToken,
     );
   }
 
-// ignore: unused_element
   AuthTokenDto fromJson(Map<String, Object> json) {
     return AuthTokenDto.fromJson(json);
   }
 }
 
 /// @nodoc
-// ignore: unused_element
 const $AuthTokenDto = _$AuthTokenDtoTearOff();
 
 /// @nodoc
 mixin _$AuthTokenDto {
-  String get accessToken;
+  String get accessToken => throw _privateConstructorUsedError;
 
-  Map<String, dynamic> toJson();
+  Map<String, dynamic> toJson() => throw _privateConstructorUsedError;
   @JsonKey(ignore: true)
-  $AuthTokenDtoCopyWith<AuthTokenDto> get copyWith;
+  $AuthTokenDtoCopyWith<AuthTokenDto> get copyWith =>
+      throw _privateConstructorUsedError;
 }
 
 /// @nodoc
@@ -60,11 +62,13 @@ class _$AuthTokenDtoCopyWithImpl<$Res> implements $AuthTokenDtoCopyWith<$Res> {
 
   @override
   $Res call({
-    Object accessToken = freezed,
+    Object? accessToken = freezed,
   }) {
     return _then(_value.copyWith(
-      accessToken:
-          accessToken == freezed ? _value.accessToken : accessToken as String,
+      accessToken: accessToken == freezed
+          ? _value.accessToken
+          : accessToken // ignore: cast_nullable_to_non_nullable
+              as String,
     ));
   }
 }
@@ -91,11 +95,13 @@ class __$AuthTokenDtoCopyWithImpl<$Res> extends _$AuthTokenDtoCopyWithImpl<$Res>
 
   @override
   $Res call({
-    Object accessToken = freezed,
+    Object? accessToken = freezed,
   }) {
     return _then(_AuthTokenDto(
-      accessToken:
-          accessToken == freezed ? _value.accessToken : accessToken as String,
+      accessToken: accessToken == freezed
+          ? _value.accessToken
+          : accessToken // ignore: cast_nullable_to_non_nullable
+              as String,
     ));
   }
 }
@@ -104,9 +110,7 @@ class __$AuthTokenDtoCopyWithImpl<$Res> extends _$AuthTokenDtoCopyWithImpl<$Res>
 
 /// @nodoc
 class _$_AuthTokenDto extends _AuthTokenDto {
-  const _$_AuthTokenDto({@required this.accessToken})
-      : assert(accessToken != null),
-        super._();
+  const _$_AuthTokenDto({required this.accessToken}) : super._();
 
   factory _$_AuthTokenDto.fromJson(Map<String, dynamic> json) =>
       _$_$_AuthTokenDtoFromJson(json);
@@ -144,15 +148,16 @@ class _$_AuthTokenDto extends _AuthTokenDto {
 }
 
 abstract class _AuthTokenDto extends AuthTokenDto {
+  const factory _AuthTokenDto({required String accessToken}) = _$_AuthTokenDto;
   const _AuthTokenDto._() : super._();
-  const factory _AuthTokenDto({@required String accessToken}) = _$_AuthTokenDto;
 
   factory _AuthTokenDto.fromJson(Map<String, dynamic> json) =
       _$_AuthTokenDto.fromJson;
 
   @override
-  String get accessToken;
+  String get accessToken => throw _privateConstructorUsedError;
   @override
   @JsonKey(ignore: true)
-  _$AuthTokenDtoCopyWith<_AuthTokenDto> get copyWith;
+  _$AuthTokenDtoCopyWith<_AuthTokenDto> get copyWith =>
+      throw _privateConstructorUsedError;
 }
