@@ -11,7 +11,7 @@ abstract class ValueObject<T> {
   /// Throws [UnexpectedValueError] containing the [ValueFailure]
   T getOrCrash() {
     // id = identity - same as writing (right) => right
-    return value.fold((f) => throw UnexpectedValueError(f), id);
+    return value.fold(((f) => throw UnexpectedValueError(f)) as T Function(ValueFailure<T>), id);
   }
 
   Either<ValueFailure<dynamic>, Unit> get failureOrUnit {
