@@ -1,5 +1,5 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
-// ignore_for_file: deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies
+// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides
 
 part of 'answer_dto.dart';
 
@@ -8,6 +8,10 @@ part of 'answer_dto.dart';
 // **************************************************************************
 
 T _$identity<T>(T value) => value;
+
+final _privateConstructorUsedError = UnsupportedError(
+    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more informations: https://github.com/rrousselGit/freezed#custom-getters-and-methods');
+
 AnswerDto _$AnswerDtoFromJson(Map<String, dynamic> json) {
   return _AnswerDto.fromJson(json);
 }
@@ -16,24 +20,21 @@ AnswerDto _$AnswerDtoFromJson(Map<String, dynamic> json) {
 class _$AnswerDtoTearOff {
   const _$AnswerDtoTearOff();
 
-// ignore: unused_element
   _AnswerDto call() {
     return const _AnswerDto();
   }
 
-// ignore: unused_element
   AnswerDto fromJson(Map<String, Object> json) {
     return AnswerDto.fromJson(json);
   }
 }
 
 /// @nodoc
-// ignore: unused_element
 const $AnswerDto = _$AnswerDtoTearOff();
 
 /// @nodoc
 mixin _$AnswerDto {
-  Map<String, dynamic> toJson();
+  Map<String, dynamic> toJson() => throw _privateConstructorUsedError;
 }
 
 /// @nodoc
@@ -97,8 +98,8 @@ class _$_AnswerDto extends _AnswerDto {
 }
 
 abstract class _AnswerDto extends AnswerDto {
-  const _AnswerDto._() : super._();
   const factory _AnswerDto() = _$_AnswerDto;
+  const _AnswerDto._() : super._();
 
   factory _AnswerDto.fromJson(Map<String, dynamic> json) =
       _$_AnswerDto.fromJson;
